# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: Learning Voice Agent API
  description: |
    AI-powered voice conversation API for learning capture and knowledge management.

    ## Overview
    The Learning Voice Agent provides multiple interfaces for capturing and retrieving
    conversational learning content:

    - **REST API**: Synchronous request/response for simple integrations
    - **WebSocket**: Real-time bidirectional streaming for audio conversations
    - **Twilio Webhooks**: Phone-based voice conversations

    ## Authentication
    Currently, the API does not require authentication for development.
    Production deployments should implement appropriate authentication mechanisms.

    ## Rate Limiting
    No rate limiting is currently enforced. Production deployments should
    implement rate limiting based on session_id or client IP.

    ## Error Handling
    All endpoints return standard HTTP status codes:
    - `200`: Success
    - `400`: Bad Request (invalid input)
    - `403`: Forbidden (invalid Twilio signature)
    - `500`: Internal Server Error
  version: 1.0.0
  contact:
    name: Learning Voice Agent Team
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.example.com
    description: Production server (configure as needed)
...
paths:
...
  /api/conversation:
    post:
      tags:
        - Conversation
      summary: Process a conversation turn
      description: |
        Handles a single conversation turn. Accepts either text input or
        base64-encoded audio. If audio is provided, it will be transcribed
        using OpenAI Whisper before processing.

        The endpoint:
        1. Transcribes audio (if provided)
        2. Retrieves conversation context from Redis
        3. Generates response using Claude Haiku
        4. Updates conversation state asynchronously
        5. Returns the response immediately
      operationId: handleConversation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ConversationRequest'
            examples:
              textInput:
                summary: Text input
                value:
                  session_id: abc-123
                  text: I'm learning about distributed systems
              audioInput:
                summary: Audio input
                value:
                  session_id: abc-123
                  audio_base64: UklGRiQAAABXQVZFZm10...
              newSession:
                summary: New session (no session_id)
                value:
                  text: Let me tell you about my new project
      responses:
        '200':
          description: Conversation response generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ConversationResponse'
              example:
                session_id: abc-123
                user_text: I'm learning about distributed systems
                agent_text: That's exciting! Distributed systems are fundamental to modern software. What aspect interests you most - consistency models, fault tolerance, or scalability patterns?
                intent: statement
                timestamp: '2024-01-15T10:30:00Z'
        '400':
          description: Bad request - no input provided
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: No input provided
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Failed to process conversation
  /api/search:
    post:
      tags:
        - Search
      summary: Search captured content
      description: |
        Performs full-text search across all captured conversations using
        SQLite FTS5 with BM25 ranking for relevance scoring.

        The search supports:
        - Natural language queries
        - Phrase matching (use quotes)
        - Boolean operators (AND, OR, NOT)
        - Prefix matching (word*)

        Results include highlighted snippets showing match context.
      operationId: searchCaptures
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SearchRequest'
            examples:
              simpleSearch:
                summary: Simple keyword search
                value:
                  query: distributed systems
                  limit: 20
              phraseSearch:
                summary: Phrase search
                value:
                  query: '"consensus algorithm"'
                  limit: 10
              booleanSearch:
                summary: Boolean search
                value:
                  query: kubernetes OR docker
                  limit: 50
      responses:
        '200':
          description: Search results returned successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SearchResponse'
              example:
                query: distributed systems
                results:
                  - id: 42
                    session_id: session-123
                    timestamp: '2024-01-15T10:30:00Z'
                    user_text: I've been studying distributed systems
                    agent_text: Great topic! What specific area?
                    user_snippet: studying <mark>distributed</mark> <mark>systems</mark>
                    agent_snippet: Great topic...
                count: 1
...
  /api/session/{session_id}/history:
    get:
      tags:
        - Sessions
      summary: Get session conversation history
      description: |
        Retrieves the conversation history for a specific session.
        Results are ordered chronologically (oldest first).
      operationId: getSessionHistory
      parameters:
        - name: session_id
          in: path
          required: true
          description: Unique session identifier
          schema:
            type: string
          example: abc-123
        - name: limit
          in: query
          required: false
          description: Maximum number of exchanges to return
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
          example: 10
      responses:
        '200':
          description: Session history retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionHistoryResponse'
              example:
                session_id: abc-123
                history:
                  - id: 1
                    timestamp: '2024-01-15T10:00:00Z'
                    user_text: Hello, I want to learn Python
                    agent_text: Great choice! Python is versatile...
                    metadata: '{"source": "api"}'
                  - id: 2
                    timestamp: '2024-01-15T10:01:00Z'
                    user_text: Where should I start?
                    agent_text: I recommend starting with basics...
                    metadata: '{"source": "api"}'
                count: 2
  /twilio/voice:
    post:
      tags:
        - Twilio
      summary: Twilio voice webhook
      description: |
        Main entry point for Twilio voice calls. Handles incoming calls and
        generates TwiML responses for voice interaction.

        This endpoint:
        1. Validates the Twilio request signature (in production)
        2. Creates or retrieves the session based on CallSid
        3. Generates appropriate TwiML based on call status
        4. Updates session metadata asynchronously

        **Note**: This endpoint expects `application/x-www-form-urlencoded` data
        from Twilio, not JSON.
      operationId: handleTwilioVoice
      requestBody:
        required: true
        content:
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TwilioVoiceRequest'
      responses:
        '200':
          description: TwiML response generated
          content:
            application/xml:
              schema:
                type: string
              example: |
                <?xml version="1.0" encoding="UTF-8"?>
                <Response>
                  <Gather input="speech" timeout="3" language="en-US"
                          action="/twilio/process-speech?session_id=twilio_CA123"
                          method="POST" speechTimeout="auto" enhanced="true">
                    <Say voice="Polly.Joanna-Neural" language="en-US">
                      Hello! I'm your learning companion. Tell me what you're working on.
                    </Say>
                  </Gather>
                  <Say>I didn't catch that. Let's try again.</Say>
                  <Redirect>/twilio/voice?session_id=twilio_CA123</Redirect>
                </Response>
        '403':
          description: Invalid Twilio request signature
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
              example:
                detail: Invalid request signature
...
components:
  schemas:
    ConversationRequest:
      type: object
      description: Request body for conversation endpoint
      properties:
        session_id:
          type: string
          description: |
            Session identifier for conversation context. If not provided,
            a new UUID will be generated.
          example: abc-123
        text:
          type: string
          description: Text input for the conversation
          example: I'm learning about distributed systems
        audio_base64:
          type: string
          description: |
            Base64-encoded audio data. Supported formats: WAV, MP3, M4A, WEBM.
            Maximum duration: 60 seconds.
          example: UklGRiQAAABXQVZFZm10...
      example:
        session_id: abc-123
        text: I'm learning about distributed systems
    ConversationResponse:
      type: object
      description: Response from conversation endpoint
      required:
        - session_id
        - user_text
        - agent_text
      properties:
        session_id:
          type: string
          description: Session identifier
          example: abc-123
        user_text:
          type: string
          description: Transcribed or provided user input
          example: I'm learning about distributed systems
        agent_text:
          type: string
          description: AI-generated response
          example: That's exciting! What aspect interests you most?
        intent:
          type: string
          description: Detected conversation intent
          enum:
            - statement
            - question
            - end_conversation
            - clarification
          default: statement
          example: statement
        timestamp:
          type: string
          format: date-time
          description: Response timestamp (UTC)
          example: '2024-01-15T10:30:00Z'
    SearchRequest:
      type: object
      description: Request body for search endpoint
      required:
        - query
      properties:
        query:
          type: string
          minLength: 1
          description: |
            Search query using FTS5 syntax. Supports:
            - Simple keywords: distributed systems
            - Phrases: "consensus algorithm"
            - Boolean: kubernetes OR docker
            - Prefix: contain*
          example: distributed systems
        limit:
          type: integer
          minimum: 1
          maximum: 100
          default: 20
          description: Maximum number of results to return
          example: 20
    SearchResponse:
      type: object
      description: Response from search endpoint
      required:
        - query
        - results
        - count
      properties:
        query:
          type: string
          description: Original search query
          example: distributed systems
        results:
          type: array
          description: Array of matching captures
          items:
            $ref: '#/components/schemas/SearchResult'
        count:
          type: integer
          description: Number of results returned
          example: 5
    SearchResult:
      type: object
      description: Individual search result
      properties:
        id:
          type: integer
          description: Capture ID
          example: 42
        session_id:
          type: string
          description: Session that captured this exchange
          example: session-123
        timestamp:
          type: string
          format: date-time
          description: When the capture was recorded
          example: '2024-01-15T10:30:00Z'
        user_text:
          type: string
          description: User's original message
          example: I've been studying distributed systems
        agent_text:
          type: string
          description: Agent's response
          example: Great topic! What specific area interests you?
        user_snippet:
          type: string
          description: User text with highlighted matches
          example: studying <mark>distributed</mark> <mark>systems</mark>
        agent_snippet:
          type: string
          description: Agent text with highlighted matches
          example: Great topic...
...
    SessionHistoryResponse:
      type: object
      description: Session history response
      properties:
        session_id:
          type: string
          description: Session identifier
          example: abc-123
        history:
          type: array
          description: Chronologically ordered exchanges
          items:
            $ref: '#/components/schemas/HistoryItem'
        count:
          type: integer
          description: Number of exchanges returned
          example: 5
...
    TwilioVoiceRequest:
      type: object
      description: Twilio voice webhook request
      properties:
        CallSid:
          type: string
          description: Unique identifier for the call
          example: CA1234567890abcdef
        From:
          type: string
          description: Caller phone number
          example: '+15551234567'
        To:
          type: string
          description: Called phone number
          example: '+15559876543'
        CallStatus:
          type: string
          description: Current call status
          enum:
            - ringing
            - in-progress
            - completed
            - busy
            - failed
            - no-answer
          example: ringing
        RecordingUrl:
          type: string
          description: URL of any recording
          example: https://api.twilio.com/...
        SpeechResult:
          type: string
          description: Transcribed speech (if using Gather)
          example: I want to learn about machine learning
...
    ErrorResponse:
      type: object
      description: Standard error response
      properties:
        detail:
          type: string
          description: Error message
          example: No input provided
...